move external_code.py for python module caching
